# server/Startup/openapi.json
{
  "x-generator": "NSwag v14.2.0.0 (NJsonSchema v11.1.0.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/Login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Login",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/Register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Register",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/Secured": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Secured",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Plants": {
      "get": {
        "tags": [
          "Plants"
        ],
        "operationId": "Plants_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlantResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Plants"
        ],
        "operationId": "Plants_Create",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlantDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Plants/{id}": {
      "get": {
        "tags": [
          "Plants"
        ],
        "operationId": "Plants_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Plants"
        ],
        "operationId": "Plants_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlantDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Plants"
        ],
        "operationId": "Plants_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/Subscribe": {
      "post": {
        "tags": [
          "Subscription"
        ],
        "operationId": "Subscription_Subscribe",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeSubscriptionDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/Unsubscribe": {
      "post": {
        "tags": [
          "Subscription"
        ],
        "operationId": "Subscription_Unsubscribe",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeSubscriptionDto"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/ExampleBroadcast": {
      "post": {
        "tags": [
          "Subscription"
        ],
        "operationId": "Subscription_ExampleBroadcast",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExampleBroadcastDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "jwt"
        ],
        "properties": {
          "jwt": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "AuthRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "minLength": 3
          },
          "password": {
            "type": "string",
            "minLength": 4
          }
        }
      },
      "PlantResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "plantName",
          "plantType",
          "moistureLevel",
          "moistureThreshold",
          "isAutoWateringEnabled"
        ],
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "plantName": {
            "type": "string",
            "minLength": 1
          },
          "plantType": {
            "type": "string",
            "minLength": 1
          },
          "moistureLevel": {
            "type": "number",
            "format": "double"
          },
          "moistureThreshold": {
            "type": "number",
            "format": "float"
          },
          "isAutoWateringEnabled": {
            "type": "boolean"
          },
          "isOwner": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
      "CreatePlantDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "plantName": {
            "type": "string"
          },
          "plantType": {
            "type": "string"
          },
          "moistureThreshold": {
            "type": "number",
            "format": "float"
          },
          "moistureLevel": {
            "type": "number",
            "format": "float"
          },
          "isAutoWateringEnabled": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          }
        }
      },
      "UpdatePlantDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "plantName": {
            "type": "string",
            "nullable": true
          },
          "plantType": {
            "type": "string",
            "nullable": true
          },
          "moistureLevel": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "moistureThreshold": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "isAutoWateringEnabled": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
      "ChangeSubscriptionDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "clientId": {
            "type": "string"
          },
          "topicIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ExampleBroadcastDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "eventType": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ApplicationBaseDto": {
        "type": "object",
        "x-abstract": true,
        "additionalProperties": false,
        "properties": {
          "eventType": {
            "type": "string"
          }
        }
      },
      "ServerBroadcastsLiveDataToDashboard": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApplicationBaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "eventType": {
                "type": "string"
              }
            }
          }
        ]
      },
      "MemberLeftNotification": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "clientId": {
                "type": "string"
              },
              "topic": {
                "type": "string"
              }
            }
          }
        ]
      },
      "BaseDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "eventType": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
      "ExampleClientDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "somethingTheClientSends": {
                "type": "string"
              }
            }
          }
        ]
      },
      "ExampleServerResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "somethingTheServerSends": {
                "type": "string"
              }
            }
          }
        ]
      },
      "ServerSendsErrorMessage": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        ]
      },
      "WaterNowClientDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "plantId": {
                "type": "string"
              }
            }
          }
        ]
      },
      "WaterNowServerResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        ]
      },
      "StringConstants": {
        "type": "string",
        "description": "Available eventType and string constants",
        "enum": [
          "ServerBroadcastsLiveDataToDashboard",
          "MemberLeftNotification",
          "ExampleClientDto",
          "ExampleServerResponse",
          "ServerSendsErrorMessage",
          "WaterNowClientDto",
          "WaterNowServerResponse",
          "Dashboard",
          "Device",
          "ChangePreferences",
          "Log"
        ]
      }
    }
  }
}